+++ CHANGELOG.md
@@ -354,6 +354,11 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Add auto refresh
 - Add runBash in `utils`
 
+
+### Added
+ - Explorer view "Enqueue" option
+ - Top Ribbon "Enqueue" button
+
 [unreleased]: https://github.com/XingYuSSS/slurm--/compare/v0.12.0...HEAD
 [0.12.0]: https://github.com/XingYuSSS/slurm--/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/XingYuSSS/slurm--/compare/v0.10.0...v0.11.0